move ml samples to e2e, make the tests standalone
