+++ samples/snippets/ml_samples/gen_ai_model_test.py
